Added confirmation request.

+++ cleanup.sh
@@ -7,31 +7,60 @@ NETWORKS=$(docker network ls | awk '{ if ($2!="bridge") print}' | awk '{ if ($2!
 if [ -z "$PS" ]; then
    echo "No Docker Processes left behind."
 else
-   echo "Killing and removing processes. Ctrl+C to abort."
-   sleep 5
-   docker kill $PS && docker rm $PS
+   echo "Targets acquired. Kill and delete?"
+   read -r -p "[y/N]" response
+   response=${response,,}
+   if [[ $response =~ ^(yes|y)$ ]]; then   
+        echo "Killing and removing processes. Ctrl+C to abort."
+        sleep 5
+        docker kill $PS
+        docker rm $PS
+   else
+        echo "Ok moving on."
+   fi
 fi
 
 if [ -z "$IMAGES" ]; then
    echo "No Images to delete."
 else
-   echo "Deleting all Images. Ctrl+C to abort."
-   sleep 5
-   docker rmi $IMAGES
+   echo "Targets acquired. Kill and delete?"
+   read -r -p "[y/N]" response
+   response=${response,,}  
+   if [[ $response =~ ^(yes|y)$ ]]; then
+        echo "Deleting all local images. Ctrl+C to abort."
+        sleep 5
+        docker rmi $IMAGES
+   else
+        echo "Ok moving on."
+   fi
 fi
 
 if [ -z "$VOLUMES" ]; then
     echo "No Docker volumes to delete."
 else
-    echo "Deleting all Docker volumes. Ctrl+C to abort."
-    sleep 5
-    docker volume rm $VOLUMES
+    echo "Targets acquired. Kill and delete?"
+    read -r -p "[y/N]" response
+    response=${response,,}    
+    if [[ $response =~ ^(yes|y)$ ]]; then
+        echo "Deleting all local docker volumes. Ctrl+C to abort."
+        sleep 5
+        docker volume rm $VOLUMES
+    else
+        echo "Ok moving on."
+    fi
 fi
 
 if [ -z "$NETWORKS" ]; then
     echo "No Networks to delete."
 else
-    echo "Deleting all Docker networks. Ctrl+C to abort."
-    sleep 5
-    docker network rm $NETWORKS
+    echo "Targets acquired. Kill and delete?"
+    read -r -p "[y/N]" response
+    response=${response,,}
+    if [[ $response =~ ^(yes|y)$ ]]; then
+        echo "Deleting all local docker networks. Ctrl+C to abort."
+        sleep 5
+        docker network rm $NETWORKS
+    else
+        echo "Ok moving on."
+    fi
 fi
